add simple bag window and fix equiping items from non-player inventory
