+++ Ozzyria.MonoGameClient/LocalState.cs
@@ -8,6 +8,7 @@ namespace Ozzyria.MonoGameClient
     internal class BagState
     {
         public uint EntityId { get; set; }
+        public string Name { get; set; }
         public List<Entity> Contents { get; set; }
         public DateTime LastSyncRequest { get; set; }
     }
@@ -22,13 +23,19 @@ namespace Ozzyria.MonoGameClient
         public uint PlayerEntityId { get; set; }
         public Dictionary<uint, BagState> BagContents { get; set; } = new Dictionary<uint, BagState>();
 
+        public bool HasBag(uint entityId)
+        {
+            return BagContents.ContainsKey(entityId);
+        }
+
         public BagState GetBag(uint entityId)
         {
             if (!BagContents.ContainsKey(entityId))
             {
-                BagContents[entityId] = new BagState
+                return new BagState
                 {
                     EntityId = entityId,
+                    Name = "",
                     Contents = new List<Entity>(),
                     LastSyncRequest = DateTime.MinValue
                 };
@@ -37,12 +44,13 @@ namespace Ozzyria.MonoGameClient
             return BagContents[entityId];
         }
 
-        public void SetBagContents(uint entityId,  List<Entity> contents)
+        public void SetBagContents(uint entityId, string name,  List<Entity> contents)
         {
             if (!BagContents.ContainsKey(entityId))
             {
                 BagContents[entityId] = new BagState{
                     EntityId = entityId,
+                    Name = name,
                     Contents = new List<Entity>(),
                     LastSyncRequest = DateTime.MinValue
                 };
@@ -50,11 +58,12 @@ namespace Ozzyria.MonoGameClient
 
             BagContents[entityId].Contents.Clear();
             BagContents[entityId].Contents.AddRange(contents);
+            BagContents[entityId].Name = name;
         }
 
         public void ForgetBagContents(uint entityId)
         {
-            if(BagContents.ContainsKey((uint)entityId))
+            if(BagContents.ContainsKey(entityId))
             {
                 BagContents.Remove(entityId);
             }
