+++ Ozzyria.MonoGameClient/UI/WindowManager.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
-using Ozzyria.MonoGameClient.UI.Windows;
 using System;
 using System.Collections.Generic;
 using System.Linq;
